+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Updated [@adguard/diff-builder] to 1.1.4.
+
+[Unreleased]: https://github.com/AdguardTeam/FiltersDownloader/compare/v2.4.2...HEAD
+
 ## [2.4.2] - 2025-06-30
 
 ### Changed